Optimized JS build: implemented Code Splitting, Tree Shaking, and Re-exports
