+++ package.json
@@ -15,6 +15,7 @@
     "@types/lodash.debounce": "^4.0.9",
     "@types/node": "^22.13.1",
     "@types/qs": "^6.9.18",
+    "@types/react-loadable": "^5.5.11",
     "ahooks": "^3.8.4",
     "axios": "^1.7.9",
     "clsx": "^2.1.1",
@@ -24,6 +25,7 @@
     "react": "^18.3.1",
     "react-content-loader": "^7.0.2",
     "react-dom": "^18.3.1",
+    "react-loadable": "^5.5.0",
     "react-paginate": "^8.3.0",
     "react-redux": "^9.2.0",
     "react-router-dom": "^6.28.2",
