+++ src/components/PizzaBlock/Skeleton.tsx
@@ -1,6 +1,6 @@
 import ContentLoader from 'react-content-loader';
 
-const Skeleton = () => (
+export const Skeleton = () => (
   <ContentLoader
     className="pizza-block"
     speed={2}
@@ -16,5 +16,3 @@ const Skeleton = () => (
     <rect x="125" y="427" rx="24" ry="24" width="152" height="45" />
   </ContentLoader>
 );
-
-export default Skeleton;
