+++ src/components/index.ts
@@ -1,0 +1,10 @@
+export * from './PizzaBlock';
+export * from './PizzaBlock/Skeleton';
+export * from './Search';
+export * from './Sort';
+export * from './Pagination';
+export * from './NotFoundBlock';
+export * from './Header';
+export * from './Categories';
+export * from './CartItem';
+export * from './CartEmpty';
