+++ src/layouts/MainLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 
-import { Header } from '../components/Header';
+import { Header } from '../components';
 import { FC } from 'react';
 
 export const MainLayout: FC = () => {
